feat: get article item with suspense
